Some tweaks after feedback from ELS reviewers.
The feedback was as follows:

Dear Francois-Rene,

This message is going out to all who submitted papers or abstract for demonstration to ELS 2014, the European Lisp Symposium.

In previous mail sometime during the last week, you should have received notification about whether or not your submission (ASDF3, or Why Lisp is Now an Acceptable Scripting Language) was accepted.  If you did not receive such notification of its disposition, please contact me immediately.

Enclosed in this message are the reviewers' comments on your submission.

Whether your submission was accepted for ELS 2014 and you are making refinements for publication, or your submission was not accepted on this go-around and you're contemplating how to improve it for possible later submission, I think you'll find some valuable insights.

You are not required to make all of the changes suggested by the various reviewers.  The suggestions, taken as a whole, may not in some cases even be self-consistent, so it may not be possible to take all suggestions. Yet I recommend that you read each of the comments with an open mind and at least seriously consider whether there is a way to use the information you have been offered to improve your work.  In the end, though, it's still your presentation and your decision.

Please feel free to contact me if you have any questions or concerns.

Thanks again for your being part of the ELS community.

 --Kent Pitman (program chair)

----------------------- REVIEW 1 ---------------------
PAPER: 9
TITLE: ASDF3, or Why Lisp is Now an Acceptable Scripting Language
AUTHORS: Francois-Rene Rideau

OVERALL EVALUATION: 2 (accept)

----------- REVIEW -----------
The paper looks like a status report rather than a technical paper.  But the contents seem interesting.

----------------------- REVIEW 2 ---------------------
PAPER: 9
TITLE: ASDF3, or Why Lisp is Now an Acceptable Scripting Language
AUTHORS: Francois-Rene Rideau

OVERALL EVALUATION: 1 (weak accept)

----------- REVIEW -----------
The paper describes the motivation, design and solutions in ASDF 3.

The design aims at keeping the system small and more modular than its predecessors.
The requirements turn out to be very difficult to achieve, since the configuration system is highly dynamic, being itself part of the system being built and having to cope with the flexibility of CL, for example the use of reader macros.

The presentation has a narrative style, describing the evolution of the system and the step backs.

It refers to the story of a bug in the original implementation, which purportedly is in a missing appendix F.

While sympathetic with the aims, I am not completely convinced of the approach.
For example, the author criticizes build systems for other languages that involve different notation and scripting languages.
But then they introduce themselves a special notation for cl-launch.
One of fundamental aspects of Lisp is the uniform notation and breaking it should be avoided.

----------------------- REVIEW 3 ---------------------
PAPER: 9
TITLE: ASDF3, or Why Lisp is Now an Acceptable Scripting Language
AUTHORS: Francois-Rene Rideau

OVERALL EVALUATION: 0 (borderline paper)

----------- REVIEW -----------
This paper describes the design and implementation issues of a major revision
(rev 3) of the Common Lisp build system ASDF. I want to like this paper, since
it's clear that the author has put an immense amount of thought and work into
this system, and clearly has some lessons to share that have come out of the
experience. But the presentation of the paper is not up to the task, and I
think it needs work. Part of the problem is that this paper is clearly a
cut-down version of a longer report, and perhaps some important things were
elided in the reduction process.

I felt the author assumed too much of the reader. I know Common Lisp pretty
well. I *don't* know ASDF, however. Unfortunately for me, the paper never
really described how ASDF works, beyond very high-level descriptions and some
small examples. For example, I was completely lost by the description of
propagation direction in section 2.11. What direction is sideways? Which way
is up and which way is down? What is it we are propagating? I know that there
are two DAGs, but I'm not sure of much more than that, and so have no idea
what's going on here. I had similar experiences all through the paper. The
discussion of one-package-per-file design, which is apparently taken from
another package manager, was totally lost on me. There's an interesting idea
here; it should be presented clearly. The paper doesn't really have a coherent
story; rather it seems to bounce around from topic to topic in a way that I
couldn't put together into a whole: we moved from bundles to shell
abstractions to sociological discussion of backwards compatibility.

The author is trying to do a lot of things in this paper.
- He's trying to talk about *lessons learned* from the experience.
- He's trying to describe little facets of the design (such as the
  shell-out subsystem).
- He's trying to discuss the architecture of the the design.
That's a lot going on -- and it all has to fit in 8 pages. What happens is
that the author winds up, in the Danish expression, "sitting down between
two stools."

In 8 pages, you've only got space to tell one story. The little facets
are the least interesting part. A lessons-learned writeup is something
that is very hard to do well; for that, I'd recommend getting a co-author
to help hash things out. OR write a great paper about the architecture of
ASDF 3, with less detailed sections listing technical facets less concretely
("It was very challenging to design an API permitting portable invocation
of programs, across multiple operating systems and filesystems."), and then,
perhaps, a cut-down lessons-learned section in the final page.

Package management / build systems are fascinating & subtle. Despite the author's kind
and even-handed words concerning the traditional farrago of Unix tools (Make, autoconf,
etc.), these non-Common-Lisp alternatives are just awful, so it's very
interesting to see what the CL world can do in counterpoint. I don't think
this paper does the author's work justice. I think what he's got is likely to
be illuminating only to people who already know some revision of ASDF. He
should produce a work that might communicate ideas to non-ASDF /
non-CommonLisp people.